% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMRHLP.R
\docType{class}
\name{ModelMRHLP-class}
\alias{ModelMRHLP-class}
\alias{ModelMRHLP}
\title{A Reference Class which represents a fitted MRHLP model.}
\description{
ModelMRHLP represents a \link[=ModelMRHLP]{MRHLP} model for which parameters have
been estimated.
}
\section{Fields}{

\describe{
\item{\code{paramMRHLP}}{A \link{ParamMRHLP} object. It contains the estimated values of the parameters.}

\item{\code{statMRHLP}}{A \link{StatMRHLP} object. It contains all the statistics associated to the MRHLP model.}
}}


\seealso{
\link{ParamMRHLP}, \link{StatMRHLP}
}
