% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatMRHLP.R
\docType{class}
\name{StatMRHLP-class}
\alias{StatMRHLP-class}
\alias{StatMRHLP}
\title{A Reference Class which contains statistics of a MRHLP model.}
\description{
StatMRHLP contains all the statistics associated to a \link[=ParamMRHLP]{MRHLP}
model.
}
\section{Fields}{

\describe{
\item{\code{pi_ik}}{Matrix of size \eqn{(m, K)} representing the prior/logistic
probabilities \eqn{\pi_{k}(x_{i}; \boldsymbol{\Psi}) = P(z_{i} = k |
  \boldsymbol{x}; \Psi)}{\pi_{k}(x_{i}; \Psi) = P(z_{i} = k | x; \Psi)} of
the latent variable \eqn{z_{i}, i = 1,\dots,m}.}

\item{\code{z_ik}}{Hard segmentation logical matrix of dimension \eqn{(m, K)}
obtained by the Maximum a posteriori (MAP) rule: \eqn{z\_ik = 1 \
  \textrm{if} \ z\_ik = \textrm{arg} \ \textrm{max}_{k} \ \pi_{k}(x_{i};
  \boldsymbol{\Psi});\ 0 \ \textrm{otherwise}}{z_ik = 1 if z_ik = arg max_k
  \pi_{k}(x_{i}; \Psi); 0 otherwise}, \eqn{k = 1,\dots,K}.}

\item{\code{klas}}{Column matrix of the labels issued from \code{z_ik}. Its elements are
\eqn{klas(i) = k}, \eqn{k = 1,\dots,K}.}

\item{\code{tau_ik}}{Matrix of size \eqn{(m, K)} giving the posterior probability
that the observation \eqn{Y_{i}} originates from the \eqn{k}-th regression
model.}

\item{\code{polynomials}}{Array of size \eqn{(m, d, K)} giving the values of the
estimated polynomial regression components.}

\item{\code{weighted_polynomials}}{Array of size \eqn{(m, d, K)} giving the values
of the estimated polynomial regression components weighted by the prior
probabilities \code{pi_ik}.}

\item{\code{Ex}}{Matrix of size \emph{(m, d)}. \code{Ex} is the curve expectation
(estimated signal): sum of the polynomial components weighted by the
logistic probabilities \code{pi_ik}.}

\item{\code{loglik}}{Numeric. Observed-data log-likelihood of the MRHLP model.}

\item{\code{com_loglik}}{Numeric. Complete-data log-likelihood of the MRHLP model.}

\item{\code{stored_loglik}}{List. Stored values of the log-likelihood at each EM
iteration.}

\item{\code{stored_com_loglik}}{List. Stored values of the Complete log-likelihood
at each EM iteration.}

\item{\code{BIC}}{Numeric. Value of BIC (Bayesian Information Criterion).}

\item{\code{ICL}}{Numeric. Value of ICL (Integrated Completed Likelihood).}

\item{\code{AIC}}{Numeric. Value of AIC (Akaike Information Criterion).}

\item{\code{log_piik_fik}}{Matrix of size \eqn{(m, K)} giving the values of the
logarithm of the joint probability \eqn{P(y_{i}, \ z_{i} = k |
  \boldsymbol{x}, \boldsymbol{\Psi})}{P(y_{i}, z_{i} = k | x, \Psi)}, \eqn{i
  = 1,\dots,m}.}

\item{\code{log_sum_piik_fik}}{Column matrix of size \emph{m} giving the values of
\eqn{\textrm{log} \sum_{k = 1}^{K} P(y_{i}, \ z_{i} = k | \boldsymbol{x},
  \boldsymbol{\Psi})}{log \sum_{k = 1}^{K} P(y_{i}, z_{i} = k | x, \Psi)},
\eqn{i = 1,\dots,m}.}

\item{\code{cpu_time}}{Numeric. Average computing time of a EM algorithm run.}
}}

\section{Methods}{

\describe{
\item{\code{EStep(paramMRHLP)}}{Method used in the EM algorithm to update statistics based on parameters
provided by \code{paramMRHLP} (prior and posterior probabilities).}

\item{\code{MAP()}}{calcule une partition d'un echantillon par la regle du Maximum A Posteriori a partir des probabilites a posteriori
Entrees : post_probas , Matrice de dimensions [n x K] des probabibiltes a posteriori (matrice de la partition floue)
n : taille de l'echantillon
K : nombres de classes
klas(i) = arg   max (post_probas(i,k)) , for all i=1,...,n
1<=k<=K
= arg   max  p(zi=k|xi;theta)
1<=k<=K
= arg   max  p(zi=k;theta)p(xi|zi=k;theta)/sum{l=1}^{K}p(zi=l;theta) p(xi|zi=l;theta)
1<=k<=K
Sorties : classes : vecteur collones contenant les classe (1:K)
Z : Matrice de dimension [nxK] de la partition dure : ses elements sont zik, avec zik=1 si xi
appartient la classe k (au sens du MAP) et zero sinon.}
}}

\seealso{
\link{ParamMRHLP}
}
