% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMRHLP.R
\docType{class}
\name{ModelMRHLP-class}
\alias{ModelMRHLP-class}
\alias{ModelMRHLP}
\title{A Reference Class which represents a fitted MRHLP model.}
\description{
ModelMRHLP represents an estimated MRHLP model.
}
\section{Fields}{

\describe{
\item{\code{paramMRHLP}}{A \link{ParamMRHLP} object. It contains the
estimated values of the parameters.}

\item{\code{statMRHLP}}{A \link{StatMRHLP} object. It contains all the
statistics associated to the MRHLP model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("regressors", "estimatedsignal"))}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"regressors" = } Polynomial regression components.
      \item \code{"estimatedsignal" = } Estimated signal (Sum of the
      polynomial components weighted by the logistic probabilities.
    }
  }
}
By default, all the above graphs are produced.}

\item{\code{summary()}}{Summary method.}
}}

\seealso{
\link{ParamMRHLP}, \link{StatMRHLP}
}
